requirements.txt
blinker==1.9.0
Bootstrap-Flask==2.5.0
click==8.1.8
dnspython==2.7.0
dominate==2.9.1
email-validator==2.3.0
Flask==3.1.2
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
python-dotenv==1.2.1
visitor==0.1.3
Werkzeug==3.1.4
WTForms==3.2.1
zipp==3.23.0
